backend/requirements.txt
# Core framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
cryptography==42.0.0

# OpenAI
openai==1.10.0

# HTTP client
httpx==0.26.0

# Environment variables
python-dotenv==1.0.0

# Date handling
python-dateutil==2.8.2

# Utilities
tenacity==8.2.3
